fix(sched): correct idle-CPU domain membership and load weight type
Address Task 4 review findings: use topo.domainOfCpu(cpu) for the
fast-path idle-CPU membership test (raw 1L<<cpu is wrong for cpu>=64),
widen RustyLoadTracker.load weight to long to match QueuedTask.weight,
and stop the class Javadoc from claiming tick() already balances (that
is wired in a later step).

+++ bpf/src/main/java/me/bechberger/ebpf/bpf/userspace/RustyLoadTracker.java
@@ -66,7 +66,7 @@ public final class RustyLoadTracker {
     }
 
     /** Decayed-at-read task load = dutyCycle * weight. */
-    public double load(int pid, int weight, long nowNs) {
+    public double load(int pid, long weight, long nowNs) {
         return dutyCycle(pid, nowNs) * weight;
     }
